Create catboost model - for data maining competition
After a series of feature engineering and de-reprocessing of the data in the database established by us, the ensemble learning CatBoost regression algorithm is used for model training. Through training, a relatively accurate prediction effect is obtained. At the same time, the K-fold verification algorithm is applied in the training process to further improve the accuracy of the model and ensure that it will not overfit. It also contains some correlation charts between features and some descriptive analysis, and finally gives a bar chart of the importance of features.
In addition, we also designed the model to obtain the predicted value of the rent price according to the user's input of relevant eigenvalues in the client, which can be used for market pricing or price detection